perf: int8_t for QuantCache sign array
Sign is only +1 or -1 — int8_t instead of int shrinks the sign
array from 2304 to 576 bytes. Combined with float pow34_sf (2304
bytes) and int16_t ix (1152 bytes), the total hot working set for
the quantize loop drops from 11,520 to 4,032 bytes — fits in L1.

+++ src/quantize.cpp
@@ -43,7 +43,7 @@ static double fast_pow34(double x) {
 struct QuantCache {
     float pow34_sf[576];  // pow34(|xr|) * sf_scale — float is sufficient
                           // (result is truncated to int 0..8191 anyway)
-    int sign[576];        // +1 or -1
+    int8_t sign[576];     // +1 or -1
 };
 
 static int find_count1_start(const int16_t* ix, int rzero) {
